requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
redis>=5.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
structlog>=23.2.0
prometheus-client>=0.19.0
httpx>=0.25.0
cryptography>=41.0.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0